Found oversight of a oversight, causing the galloooping
Originally I had forgotten to convert the x axis control of dest
offset from pixel dimension to bytes dimension.

But then when I did it, not sure what I was thinking, bcas, I had
multiplied by 8, instead of 4 for the conversion, so as a matter
of fact, the tiles where gallopping on the screen.

This has been fixed, and the previously added debug print, boxed
inside a ifdef