master: Update the helper script and README
Wrt the performance check commands

In a previous update, there was a oversight wrt fbdetile video filter
test command, because of which the filter was not getting used. This
has been fixed now.

The new Performance counter based data is also shown now.

The need for using bgr0 format when using kmsgrab and hwdownload, on
intel linux machines had been missed out in the example, this has been
fixed now.