Generic based Intel Tile-X detiling
To cross verify at a simple level easily that the generic detile
logic is ok to some extent, I have implemented the Tile-X detiling
using the generic detile logic now, by setting up the required
arguments to the generic detile function suitably.

With this it is noticed that the detiling is occuring properly
for a Tile-X layout based framebuffers.

A new type called intelgx added to trigger this.