Added Generic TileX detile to time_ffmpeg mix
With this we can check the difference in processing load between
the default targeted TileX detile logic and the generic detile
logic configured for TileX.

With this the additional loading was found to be only 3% when
checked using processor performance counters.

Also added test run in time_ffmpeg for the new Generic Detile based
TileY logic, which will be added next.