Parallel detiling need tile boundry pos in dirChangeList
Parallel detiling logic requires the posOffset corresponding to
the last subtile row within the tile in the dirChangeList, so that
it knows when to adjust variables wrt the parallel detiling.

For TileX, this doesnt matter as the tile and subtile are one and
the same.

However for TileY, a new entry requires to be added, as each tile
is made up of multiple subtiles and because the subtile dirChange
entry was originally good enough for tile dirChange also. But now
because the tile boundry needs to be passed as part of dirChangeList
the same has been added into it.

NOTE: It could have been passed as a seperate argument to the detile
function, but then it will have to be checked wrt cSTR independent
of the dirChangeList, and that would make it inefficient at one level,
so added to dirChangeList itself.