Avoid x86intrinisc to compile on all architectures
x86intrinsic _rdtsc is used to get a rough idea of the time taken
by the detiling logic, by looking at the x86 perfromance counter.

However this is not required for the normal running of the logic.

So disabling the same so that it can be compiled and run on other
architectures.