kmsgrab update format_modifier, only if user didnt
Changed the kmsgrab to set the default value for format modifier
to MOD_INVALID (from the earlier MOD_NONE), when starting.

In turn user can set it to any valid value from the commandline,
if required.

This allows the read_headers logic to check if the current stored
format_modifier is the invalid default from kmsgrab or a specific
value set by the user. And only if the user has not explicitly
specified a format_modifier, update it from the value got using
GetFB2.