fbtile checkpixformats, hwdownload if no fbdetile
Added a checkpixformats function to fbtile, to verify if the given
pix formats are supported by fbtile (inturn detile) currently.

This is used as part of the hwdownload's input and output configuration
handshake, to ensure that it selects a pix format supported by detile.

In vf_fbdetile, this is not required, as it checks for this on its own.

Also hwdownload, updated to not unnecessarily create a additional output
buffer and inturn call into detile_this, if fbdetile is not enabled.

Also a fix wrt print of 64bit integer.