fbdetile videofilter cpu based framebuffer detiling
This adds a video filter called fbdetile, which allows the user
to detile framebuffer layout into a linear layout, if required.

It uses the fbtile helper routines to achieve the detiling.

This is useful, if

a) the user doesnt want to apply detiling when capturing some
content/framebuffer which is tiled. OR

b) the user already has tiled content with them. OR

c) a developer wants to experiment with tiled data.