Log once TILE_AUTO unknown/unsupported format_modifer
If a user is trying to capture screen/framebuffer from a drm hw
context (ex using kmsgrab), and inturn if the format_modifier is
not supported by fbtile, it could have led to the unsupported
warning message getting printed multiple times.

This has been fixed now. Now it logs the situation as a warning
once and as a verbose data for the remaining times in a given
session.