hwcontext_drm if detiled set format_modifier linear
If hwcontext_drm successfully applies detiling of a tiled layout,
then it changes the format_modifier of the associated hw frame to
DRM_FORMAT_MOD_LINEAR, so that its users know that hwcontext_drm
has transfered a linear layout of the tiled frame.

Ideally a normal AVFrame should also contain the format_modifer,
but because it currently doesn't so reusing the corresponding
hwAVFrame to mark this.

Resequenced check for TILE_NONE in detile_this, so that if TILE_AUTO
identifies a LINEAR layout, then the logic can return immidiately.