DetileGeneric OptiOrSimple, decide in fbtile.h
Fixed my unneeded confusion from a earlier commit today as well as
zeroed in on the reason why detile_generic was behaving like simple
mode instead of opti mode.

As the mapping of detile_generic to simple or opti occurs in the
fbtile.h file and the actual call based on same occurs in fbtile.c,
so what ever is the mapping when fbtile.c is compiled into a library
is what will persist and not any change done to mapping in one of the
users of fbtile.

And by default, the mapping which was triggering when fbtile.c was
getting compiled was the simple version and not the opti version,
so also the slightly higher TSCCnt for TileYf detile_generic.

So have removed the ineffective setting of DETILE_GENERIC_OPTI in
vf_fbdetile.c. And forced the define to OPTI in fbtile.h

Also re-added the performance figures for Tile-Yf DetileGenericOpti
mode.