fbtile_checkpixformats return consistent with fbtile returns
Previously fbtile_checkpixformats was returning 1 if given pix
formats was supported and 0 otherwise.

This was different to the convention followed by detile logic,
which was returning 0 on success and 1 on error.

So have modified checkpixformat to also send 0 on success and 1 on
error.

TODO: Fix its users to follow the new convention