fbtiler_generic, check cond outside inner loop
As _tile_generic_simple and family has been modified to be even
more generic and allow both tiling and detiling, so renamed it to
_fbtiler_generic_simple and family.

To see how the caching of code and data and branch predictions and
out of order execution and register renaming and ... impact condition
check inside and outside the inner most loop, tried moving the
condition check to outside the inner loop, but it doesnt seem to have
helped, or rather it might have become slightly more slower? or rather
at one level the differences could be noise for the given granularity
of underlying counters and other latencies including scheduling.