Provide compile time option to unroll opti inner loop
As the impact of unrolling is only small, and as it makes working
with subtiles smaller than 4x4 not possible in most situations,
its not enabled by default.

NOTE: If directly using simd or rep movs (as the case may be) will
help or not needs to be checked once. Also prefetch hinting needs
to be checked once.