Removed fbtiler_generic src const
As current generic logic, dynamically maps src and dst into either
tiled or linear buffer pointer, based on the operation being performed
and inturn as tiled and linear buffer pointers are non const, because
depending on the operation either of them can become the dest buffer.

Also cleaned up the comments a bit.