fbtilelayoutid_from_drmformatmodifier
Added id to the name, to make it clear that what we are doing by
calling this function is to get the equivalent fbtilelayout id
used to represent the specified drm format_modifier. This id is
used by fbtile logic to identify the tile layout of the framebuffer
it is working with.