FBTILE_SCOPE_LIMITED boxing of video filters
Renamed SCOPE_LIMITED to FBTILE_SCOPE_LIMITED, so that its namespaced
in a better way.

Moved the structs and mapping of fbtiler_generic in fbtile.h to be
available always irrespective of scope is public or limited.

Used FBTILE_SCOPE_LIMITED to box the fbtile helpers, so that they
can be compiled irrespecitive of whether the LIMITED scope is defined
or not. NOTE: However currently it just allows code to be compiled,
but the logic doesnt work.