hwcontext_drm: detile from hw to avframe
Previosly I had changed the original copy logic with the new copy
with tiling operation, so that one can do either tiling or detiling
as required based on the source and the destination, even thou at
that time the logic was only supporting detile.

Now fbtile has been updated to support both tiling as well as detiling
however because there is no tiling layout info in AVFrame, it is
difficult to tell, whether it currently contains a linear layout data
or a tiled layout data.

So for now I trigger the fbtiler_conv only when copying from hw to
avframe (because here I can be almost sure that the source hw is
potentially tiled, by looking at the format modifier of the source
hw framebuffer).

However for the case of copying/transferring from avframe into hw,
I no longer call the fbtiler_conv, because as mentioned above, I cant
be very sure if the user is passing me tiled or untiled data in the
avframe.

LATER if required, one can add support to auto tile from linear layout
if the tile layout of destination hw framebuffer is tiled. THis will
require a helper function similar to the drm_frame_transfer_with_detile,
but obviously called drm_frame_transfer_with_tile, and with suitable
minimal modification to the logic, to achieve tiling. Also how user
will use or more importantly bypass this tiling (if they dont want it)
needs to be thought through once. NOTE: The underlying fbtiler_conv
supports both tiling and detiling.

Or rather if AVFrame is updated to store tile layout info, in the future,
then during both transfer from hw and transfer to hw context, one may do
either tiling or detiling as the case may be. In which case a new helper
built on top of the av_frame_copy_with_tiling which I have already provided
needs to be done.