fbtile_framecopy additional status TILECOPY or COPYONLY
Added a additional status argument, which tells if both tile/detile
and copy was carried out or only a copy was carried out.

Because the function return value is shared with values returned by
av_frame_copy, so this additional status argument was added, so that
they dont clash with one another.