Fix configure wrt kmsgrab and GetFB2
require_pkg_config will exit configure, if the check fails. While
we want to only check if drmModeGetFB2 function is available or not.
So have changed to check_pkg_config.

Way things are setup in libdrm of ubuntu 16.04 or rather in its
xf86drmMode.h is such that, unless we also include xf86drm.h or
stdint.h before including xf86drmMode.h, it will fail because of
undefined uint32_t etc family of standard int definitions. So this
will fail the drmModeGetFB2 check, for unrelated reasons. So have
added xf86drm.h to the list of headers for check_pkg_config.

check_pkg_config takes care of disabling the name passed to it
and then inturn if the check is successful, then enable the name.

    Previously I was wrongly passing libdrm as the name argument,
    so if drmModeGetFB2 fails, even libdrm was getting disabled.
    This has been fixed now by passing drm_getfb2 has the name.

    These also mean that there is no need to call enable drm_getfb2
    explicilty. This has been removed now.