Direct fbtile_generic_opti|simple, opti unrolled
fbtile_generic_opti inner loop is unrolled now. However as noted
in the comment, if the subtile or walk is not a multiple of 4, you
cant use this function, in such cases use fbtile_generic_simple.

Also instead of the indirection through fbtile_generic, now directly
call fbtile_generic_opti or fbtile_generic_simple as required.