fbtile: getlayoutid
Changed from fbtilelayoutid_from_drmformatmodifier to be a more
generic logic now called fbtile_getlayoutid. It takes a family id
and the tile layout as specified by that family of code and inturn
converts it to the fbtile's equivalent tile layout id.

So in future, if fbtile is used in non drm context, basic skeletal
code is already available to map from that new context into fbtile.

hwcontext_drm updated to use this new api.