fbtile: blunt Dont tile/detile beyond given height - p1
Sometimes the framebuffer height we have to work with may not be a
multiple of the tile height.

And this cant be handled safely in a simple, clean and efficient
manner when tiling/detiling, so for now just bluntly (ie in a brute
force or non-fine way) ensure that we dont go into the last row of
tiles of the framebuffer, which has actually got tiles which are
smaller than the actual tile height.

NOTE:TOCHECK: Have to check if the hardware tilers in GPUs actually
manage such partial height (as well as partial width) tiles and or
they extend the framebuffer such that along both horizontal and
vertical axis there are full size tiles.

NOTE:TOCHECK: Even if hardware tilers extend framebuffer to ensure
that it is made up of full size tiles, does the os and ffmpeg code
inturn pass such extended framebuffers to allow proper handling of
such framebuffers.

NOTE:TODO: If hardware tilers dont extend framebuffers to ensure
that it is made up of full size tiles, and instead if they have
special logic to handle partial tiles, then at a later stage, I
will have to update the logic to handle such partial tiles. But
then in software this will be bit inefficient. In such a situation
the current blind bypassing of the last row of partial tiles is
a simple blunt way of avoiding it, at the expense of loss of few
lines of pixels at the bottom. However if tile is partial along
horizontal axis (i.e width), have to think how to handle it bit
more deeply, at that later stage, especially if proper stride /
pitch / linesize info is not available for such tiled layout.

Currently this blunt bypassing is achieved by adjusting the total
number of subtile lines one is working with.