fbtile: blunt Dont tile/detile beyond given height - p2
Now even use a linear layout current vertical position based height
check to restrict going beyond the blunt safe height calculated, i.e
in case the height of the framebuffer is not a multiple of tileHeight

THis adds a additional check on top of the tweaked
	curSubTileLine < adjustTotalSubTileLines check, added in
the previous commit.

So for one reason if one of them fail, the other will still be there
to rein things in.